refactor: 이벤트 핸들러를 ui단에서 분리
